% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ttest.R
\name{dcor.ttest}
\alias{dcor.ttest}
\title{Distance correlation T-test
It uses the result of U-statistic distance correlation to test independence for
high dimensional data}
\usage{
dcor.ttest(x, y)
}
\arguments{
\item{x}{data of x}

\item{y}{data of y}
}
\description{
Distance correlation T-test
It uses the result of U-statistic distance correlation to test independence for
high dimensional data
}
\examples{
n = 200
x = rnorm(n)
y = rnorm(n)
res = dcor.ttest(x,y)
}
