% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcov.R
\name{pcov}
\alias{pcov}
\title{Projection covariance between two random vectors
This function implements the projection correlation in \cite{Zhu et al.(2017)}}
\usage{
pcov(x, y)
}
\arguments{
\item{x}{the matrix of x}

\item{y}{the matrix of y}
}
\description{
Projection covariance between two random vectors
This function implements the projection correlation in \cite{Zhu et al.(2017)}
}
\examples{
x = matrix(rnorm(200),100,2)
y = matrix(rnorm(200),100,2)
pcov(x,y)
}
\references{
Zhu, L., Xu, K., Li, R., & Zhong, W. (2017). Projection correlation between two random vectors. Biometrika, 104(4), 829-843.
}
